requirements.txt
Django==5.0.4

django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
djoser==2.2.2
social-auth-app-django==5.4.0

django-ckeditor==6.7.1

pytz==2024.1

Pillow==10.3.0
mutagen==1.47.0

psycopg2==2.9.9 

argon2-cffi==23.1.0 
django-environ==0.11.2

stripe==9.2.0
braintree==4.28.0

pyJWT==2.1.0

boto3==1.35.9
django-storages==1.14.2 

markdown==3.6
gunicorn==22.0.0
whitenoise==6.6.0

requests==2.31.0

python-decouple==3.1